+++ src/CMakeLists.txt
@@ -17,9 +17,9 @@ target_sources(sylvan
     sylvan_stats.c
     sylvan_table.c
     sylvan_zdd.c
-    sylvan_var_swap.c
-    sylvan_var_level.c
-    sylvan_var_reorder.c
+        sylvan_varswap.c
+        sylvan_levels.c
+        sylvan_reorder.c
   PUBLIC
     sylvan.h
     sylvan_bdd.h
@@ -39,9 +39,9 @@ target_sources(sylvan
     sylvan_tls.h
     sylvan_zdd.h
     sylvan_zdd_int.h
-    sylvan_var_swap.h
-    sylvan_var_level.h
-    sylvan_var_reorder.h
+        sylvan_varswap.h
+        sylvan_levels.h
+        sylvan_reorder.h
 )
 
 set_target_properties(sylvan PROPERTIES VERSION ${sylvan_VERSION} SOVERSION ${sylvan_VERSION_MAJOR})
