+++ src/sylvan.h
@@ -59,9 +59,9 @@ namespace sylvan {
 #include <sylvan_mtbdd.h>
 #include <sylvan_bdd.h>
 #include <sylvan_ldd.h>
-#include <sylvan_var_level.h>
-#include <sylvan_var_swap.h>
-#include <sylvan_var_reorder.h>
+#include <sylvan_levels.h>
+#include <sylvan_varswap.h>
+#include <sylvan_reorder.h>
 #include <sylvan_zdd.h>
 
 #ifdef __cplusplus
