+++ src/sylvan_cache.c
@@ -30,7 +30,7 @@
 #endif
 
 #ifndef compiler_barrier
-#define compiler_barrier() { asm volatile("" ::: "memory"); }
+#define compiler_barrier() { asm volatile("dmb ish" ::: "memory"); }
 #endif
 
 #ifndef cas
