+++ src/sylvan_reorder.c
@@ -15,8 +15,8 @@
  */
 
 #include <sylvan_int.h>
-#include "sylvan_var_swap.h"
-#include "sylvan_var_level.h"
+#include "sylvan_varswap.h"
+#include "sylvan_levels.h"
 
 /*
  * TODO
