+++ test/test_basic.c
@@ -12,8 +12,8 @@
 #include "sylvan.h"
 #include "test_assert.h"
 #include "sylvan_int.h"
-#include "sylvan_var_swap.h"
-#include "sylvan_var_level.h"
+#include "sylvan_varswap.h"
+#include "sylvan_levels.h"
 #include "common.h"
